+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is simplified version of [Keep a Changelog](https://keepachangelog.co
 - Improve Local Source when loading from `android/data` (@lalalasupa0)
 - Refresh available extensions list when an extension repo is added or removed
 - Replace filter FAB with Floating Toolbar when browsing source
+- Show FAB button to read/resume chapter when start/continue reading button is off-screen
 
 ### Fixes
 - Allow users to bypass onboarding's permission step if Shizuku is installed